Finished Lotto.java

+++ AE/Java/12.04.13/Wert.java
@@ -1,9 +1,9 @@
 public class Wert{
 
 	public static void main(String[] args) {
-		int value 			= 1000;
+		double value 			= 1000;
 		int years			= 0;
-		double  value_shortage 	= 0.15;
+		double value_shortage 	= 0.15;
 
 		for(int i=0; i<years; i++){
 			value = (value-(value*(1-value_shortage)));
